+++ src/TaskException.php
@@ -1,5 +1,11 @@
 <?php
 
+/**
+ * This file is part of the Inspectioneering TaskRunner package.
+ *
+ * @license MIT License
+ */
+
 namespace Inspectioneering\TaskRunner;
 
 class TaskException extends \Exception